backend/requirements.txt
annotated-types==0.7.0
anyio==4.11.0
backports.asyncio.runner==1.2.0
certifi==2025.8.3
charset-normalizer==3.4.3
click==8.3.0
distro==1.9.0
dnspython==2.8.0
exceptiongroup==1.3.0
fastapi==0.117.1
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.10
iniconfig==2.1.0
jiter==0.11.0
lark==1.3.0
numpy==2.2.6
openai==1.109.1
packaging==25.0
pandas==2.3.2
pluggy==1.6.0
pydantic==2.11.9
pydantic_core==2.33.2
Pygments==2.19.2
pymongo==4.15.1
pytest==8.4.2
pytest-asyncio==1.2.0
python-dateutil==2.9.0.post0
python-dotenv==1.1.1
pytz==2025.2
PyYAML==6.0.3
requests==2.32.5
six==1.17.0
sniffio==1.3.1
starlette==0.48.0
tomli==2.2.1
tqdm==4.67.1
typing-inspection==0.4.1
typing_extensions==4.15.0
tzdata==2025.2
urllib3==2.5.0
uvicorn==0.37.0
